+++ apps/api/src/database/Exercises.ts
@@ -1,22 +1,23 @@
-import mysql from 'mysql2'
+import pg from 'pg'
+const { Pool } = pg
 import dotenv from 'dotenv'
-
 dotenv.config()
 
-const pool = mysql
-  .createPool({
-    host: process.env.MYSQL_HOST,
-    user: process.env.MYSQL_USER,
-    password: process.env.MYSQL_PASSWORD,
-    database: process.env.MYSQL_DATABASE,
-  })
-  .promise()
+const pool = new Pool({
+  host: process.env.POSTGRES_HOST,
+  user: process.env.POSTGRES_USER,
+  password: process.env.POSTGRES_PASSWORD,
+  database: process.env.POSTGRES_DATABASE,
+  port: parseInt(process.env.POSTGRES_PORT),
+})
+
+pool.connect()
 
 export async function getAllExercises() {
-  const [rows] = await pool.query(`SELECT * FROM exercises`)
+  const { rows } = await pool.query(`SELECT * FROM exercises`)
   return rows
 }
 
 export async function createExercise(name: string) {
-  await pool.query('INSERT INTO `exercises` (`name`) VALUES (?)', [name])
+  await pool.query(`INSERT INTO exercises (name) VALUES ($1)`, [name])
 }